# documention/swagger/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/national/user/profile": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/user/add": {
      "post": {
        "operationId": "UserController_addUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }, { "api_key": [] }]
      }
    },
    "/national/user/register": {
      "post": {
        "operationId": "UserController_registerUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/user/update": {
      "put": {
        "operationId": "UserController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserUpdateDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/user/resetPassword": {
      "put": {
        "operationId": "UserController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/PasswordUpdateDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/user/regenerateApiKey": {
      "put": {
        "operationId": "UserController_resetApiKey",
        "parameters": [
          {
            "name": "email",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/user/query": {
      "post": {
        "operationId": "UserController_queryUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/user/download": {
      "post": {
        "operationId": "UserController_getDownload",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DataExportQueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/user/delete": {
      "delete": {
        "operationId": "UserController_deleteUser",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["User"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
...
    "/national/Settings/update": {
      "post": {
        "operationId": "SettingsController_updateSettings",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SettingsDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Settings"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/Settings/query": {
      "get": {
        "operationId": "SettingsController_getSettings",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": { "type": "number" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Settings"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/Settings/signs/update": {
      "post": {
        "operationId": "SettingsController_updateSLCFSigns",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/SLCFSignsDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Settings"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/Settings/certificates": {
      "get": {
        "operationId": "SettingsController_generatePreviewCertificates",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Settings"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/national/projectManagement/query": {
      "post": {
        "operationId": "ProjectManagementController_getAll",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 2 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/projectManagement/getProjectById": {
      "post": {
        "operationId": "ProjectManagementController_getProjectById",
        "parameters": [],
        "responses": { "201": { "description": "" } },
        "tags": ["Phase 2 Project Management"],
        "security": [{ "bearer": [] }, { "bearer": [] }]
      }
    },
    "/national/projectManagement/logs": {
      "get": {
        "operationId": "ProjectManagementController_getLogs",
        "parameters": [
          {
            "name": "refId",
            "required": true,
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Phase 2 Project Management"],
        "security": [{ "bearer": [] }]
      }
    },
...
    "/national/reportsManagement/queryAefRecords": {
      "post": {
        "operationId": "ReportsManagementController_queryAefRecords",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/QueryDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Reports Management"],
        "security": [{ "bearer": [] }]
      }
    },
    "/national/reportsManagement/downloadAefReport": {
      "post": {
        "operationId": "ReportsManagementController_downloadAefReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/AefExportDto" }
            }
          }
        },
        "responses": { "201": { "description": "" } },
        "tags": ["Reports Management"],
        "security": [{ "bearer": [] }]
      }
    }
  },
  "info": {
    "title": "National API",
    "description": "RESTful Web API Documentation.",
    "version": "0.5",
    "contact": {}
  },
...
  "servers": [{ "url": "/" }],
  "components": {
    "securitySchemes": {
      "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" },
      "api_key": { "type": "apiKey", "in": "header", "name": "api_key" }
    },
    "schemas": {
      "OrganisationDto": {
        "type": "object",
        "properties": {
          "taxId": { "type": "string" },
          "paymentId": { "type": "string" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "regions": { "type": "array", "items": { "type": "string" } },
          "provinces": { "type": "array", "items": { "type": "string" } },
          "phoneNo": { "type": "string" },
          "faxNo": { "type": "string" },
          "website": { "type": "string" },
          "address": { "type": "string" },
          "logo": { "type": "string" },
          "country": { "type": "string" },
          "companyRole": {
            "type": "string",
            "enum": [
              "IC",
              "PD",
              "API",
              "DNA",
              "Ministry",
              "ClimateFund",
              "ExecutiveCommittee"
            ]
          },
          "nameOfMinister": { "type": "string" },
          "govDep": {
            "type": "string",
            "enum": [
              "1",
              "2",
              "3",
              "4",
              "5",
              "6",
              "7",
              "8",
              "9",
              "10",
              "11",
              "12",
              "13",
              "14",
              "15",
              "16",
              "17",
              "18",
              "19",
              "20",
              "21",
              "22",
              "23",
              "24",
              "25",
              "26",
              "27",
              "28",
              "29",
              "30",
              "31",
              "32",
              "33",
              "34",
              "35",
              "36",
              "37",
              "38",
              "39",
              "40",
              "41",
              "42",
              "43",
              "44",
              "45",
              "46",
              "47",
              "48",
              "49",
              "50",
              "51",
              "52",
              "53",
              "54",
              "55",
              "56",
              "115",
              "57",
              "58",
              "59",
              "60",
              "61",
              "62",
              "63",
              "64",
              "65",
              "66",
              "67",
              "68",
              "69",
              "70",
              "71",
              "72",
              "73",
              "74",
              "76",
              "77",
              "78",
              "79",
              "80",
              "81",
              "82",
              "83",
              "84",
              "85",
              "86",
              "87",
              "88",
              "89",
              "90",
              "91",
              "92",
              "93",
              "94",
              "95",
              "96",
              "97",
              "98",
              "99",
              "100",
              "101",
              "102",
              "103",
              "104",
              "105",
              "106",
              "108",
              "109",
              "110",
              "111",
              "112",
              "113",
              "114",
              "116",
              "117",
              "118",
              "119",
              "120",
              "121",
              "122",
              "123",
              "124",
              "125",
              "126",
              "127",
              "128",
              "129",
              "130",
              "131",
              "132",
              "133",
              "134",
              "135",
              "136",
              "137",
              "138",
              "139",
              "140",
              "141",
              "142",
              "143",
              "144",
              "145",
              "146",
              "147",
              "148",
              "149",
              "150",
              "151",
              "152",
              "153",
              "154",
              "155",
              "156",
              "157",
              "158",
              "159"
            ]
          },
          "ministry": {
            "type": "string",
            "enum": [
              "Agriculture",
              "Aviation",
              "Communications",
              "Economy",
              "Education",
              "Energy",
              "Environment",
              "Health",
              "Intelligence",
              "Judiciary",
              "Maritime",
              "Media",
              "Science And Technology",
              "Water Resources",
              "Other"
            ]
          },
          "sectoralScope": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "1",
                "2",
                "3",
                "4",
                "5",
                "6",
                "7",
                "8",
                "9",
                "10",
                "11",
                "12",
                "13",
                "14",
                "15"
              ]
            }
          }
        },
        "required": [
          "taxId",
          "paymentId",
          "name",
          "email",
          "companyRole",
          "nameOfMinister",
          "govDep",
          "ministry",
          "sectoralScope"
        ]
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "role": {
            "type": "string",
            "enum": ["Root", "Admin", "Manager", "ViewOnly"]
          },
          "name": { "type": "string" },
          "phoneNo": { "type": "string" },
          "company": { "$ref": "#/components/schemas/OrganisationDto" },
          "companyId": { "type": "number" }
        },
        "required": ["email", "role", "name"]
      },
      "UserUpdateDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "name": { "type": "string" },
          "phoneNo": { "type": "string" }
        },
        "required": ["id", "name"]
      },
      "PasswordUpdateDto": {
        "type": "object",
        "properties": {
          "newPassword": { "type": "string" },
          "oldPassword": { "type": "string" }
        },
        "required": ["newPassword", "oldPassword"]
      },
      "QueryDto": {
        "type": "object",
        "properties": {
          "page": { "type": "number" },
          "size": { "type": "number" },
          "filterAnd": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "filterOr": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "sort": {
            "type": "object",
            "example": { "key": "name", "order": "asc" },
            "items": { "$ref": "#/components/schemas/SortEntry" }
          },
          "filterBy": {
            "type": "object",
            "example": { "key": "ministryLevel", "values": ["1", "2"] },
            "items": { "$ref": "#/components/schemas/FilterBy" }
          },
          "extendedProperties": {
            "type": "object",
            "example": { "key": "isGetInvestmentHistory", "values": true }
          }
        },
        "required": ["page", "size"]
      },
      "DataExportQueryDto": {
        "type": "object",
        "properties": {
          "filterAnd": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "filterOr": {
            "type": "array",
            "example": [{ "key": "age", "operation": "gt", "value": 25 }],
            "items": { "$ref": "#/components/schemas/FilterEntry" }
          },
          "sort": {
            "type": "object",
            "example": { "key": "name", "order": "asc" },
            "items": { "$ref": "#/components/schemas/SortEntry" }
          },
          "filterBy": {
            "type": "object",
            "example": { "key": "ministryLevel", "values": ["1", "2"] },
            "items": { "$ref": "#/components/schemas/FilterBy" }
          }
        }
      },
...
      "SettingsDto": {
        "type": "object",
        "properties": {
          "id": { "type": "number" },
          "settingValue": { "type": "string" }
        },
        "required": ["id", "settingValue"]
      },
      "SLCFSignsDto": {
        "type": "object",
        "properties": {
          "ceoSign": { "type": "string" },
          "chairmanSign": { "type": "string" }
        }
      },
...
      "AefExportDto": {
        "type": "object",
        "properties": {
          "reportType": { "type": "string", "enum": ["HOLDINGS", "ACTIONS"] },
          "fileType": { "type": "string", "enum": ["xlsx", "csv"] }
        },
        "required": ["reportType", "fileType"]
      }
    }
  }
}